backend/requirements.txt
Django==4.2.5
djangorestframework==3.12.4
django-filter==21.1
django_extra_fields==3.0.2
python-dotenv==1.0.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.3.0